Salvar 30-05/requirements.txt
requests
pyttsx3
SpeechRecognition
pyaudio
scikit-learn
joblib
python-dotenv